Updataing Object State Immutably.
When user click on square button then put 'x' on that button.
So we need to change value null--->to--->'X'. so we need to create first of all deep copy of array. then store in new variable and update the value in new variable and then return the new variabel.