CSS Module - for scopping style with in component - using classes name.
CSS Modules is a popular approach for styling React components. It allows for scoped CSS, meaning that the styles defined in a CSS Module only apply to the specific component it is imported into. This helps to avoid naming conflicts and makes it easier to manage styles in large projects.

In CSS Modules, each CSS file is treated as a separate module. Class names are automatically generated and scoped locally to the component, ensuring that styles do not leak or clash with other components. These generated class names can be accessed in the JavaScript code using an import statement.

By importing the CSS Module in the component file, we can use the generated class names to apply styles to the corresponding elements. This provides a more modular and maintainable way of styling components, as the styles are encapsulated within the component itself.

- CSS Module only work with class name.
If you specfy global element/Tag then that styling is also applied to other componet. So, use classes when you want to apply styling with in component only.

SS Modules work by scoping class names. When you use CSS Modules, each class name in your CSS file gets transformed into a unique identifier that includes the filename and the class name. This way, the styles you define for a class in one component won't affect elements with the same class name in another component.

However, it's important to note that global styles and tag selectors are not scoped. This means that if you define styles for a tag (like h1, p, etc.) or a global class in a CSS Module, those styles will apply across your entire application, not just the component where the CSS Module is imported.