Creating flexible component with styled component.
all props to the styled component is automaticlly forwared to elemen.

in styled-components, all unknown props are automatically forwarded to the DOM element. This means that if you pass a prop that isn't explicitly handled by the styled component, it will be passed on to the rendered DOM element.

For example, in your code, the type, className, and onChange props are all passed to the Input styled component. These props are not handled by the styled component itself, so they are forwarded to the input element in the DOM.

However, it's important to note that this only applies to DOM elements. If you're styling a React component, the props are not automatically forwarded. You would need to explicitly pass the props to the underlying DOM element in the component.

Also, be aware that React has a list of reserved props (like key, ref, etc.) that are not forwarded, as they have special meanings in React.