Add Side Effect(Asking location permission) & use useEffect (for after taking permission sort the places by distances.)
- Side Effect & useEffect.
a side effect is any operation that interacts with the outside of its scope. This includes data fetching, subscriptions, timers, logging, and manually changing the DOM. React provides the useEffect hook to handle these side effects in function components. The useEffect hook runs after every render by default, but you can control when it runs by passing an array of dependencies as the second argument.

- The useEffect Hook allows you to perform side effects in your components.
- Some examples of side effects are: fetching data, directly updating the DOM, and timers.
- useEffect accepts two arguments. The second argument is optional.
- useEffect(<function>, <dependency>)