Using Redux Data in React Components.
useSelector is a hook provided by the react-redux library. It allows you to extract data from the Redux store state, using a selector function.

Why we use it: We use useSelector to access the state in the Redux store from our components. This allows us to decouple our components from the state management and makes them more reusable and easier to test.

- How to use it:The useSelector hook takes a single argument: a selector function. This function has the entire Redux store state as its argument, and should return the piece of state this component needs.

const counter = useSelector(state => state.counter);

In this example, the selector function is (state) => state.counter, which means it will extract the counter value from the state.

When to use it: You should use useSelector whenever you need to access data from the Redux store in a functional component. It's a part of the React-Redux hooks API, which is a more modern approach to accessing store data than the connect higher-order component.