Using an Action Creator Thunk.
a thunk is a middleware that allows you to write action creators that return a function instead of an action. The thunk can be used to delay the dispatch of an action, or to dispatch only if a certain condition is met. This functionality makes it possible to handle complex logic in your action creators.